blog/requirements.txt
Django==4.1.7
django-allauth==0.52.0
django-filter==22.1
djangorestframework==3.14.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3

